--- src/openms/include/OpenMS/MATH/MISC/NNLS/README
@@ -1,10 +1,0 @@
-In order to solve non-negative least square problems, we rely on C++ code
-from
-
-http://www.kyb.tuebingen.mpg.de/bs/people/suvrit/work/progs/nnls.html
-
-which is placed here unchanged.
-To keep the namesspace clear we include the .cc file in our .cpp file (which goes into the OpenMSlib).
-
-
-So *no* Makefile is required for this directory!
