+++ src/openms/includes.cmake
@@ -6,7 +6,7 @@ include(source/CONCEPT/sources.cmake)
 include(source/SYSTEM/sources.cmake)
 include(source/MATH/STATISTICS/sources.cmake)
 include(source/MATH/MISC/sources.cmake)
-include(source/MATH/MISC/NNLS/sources.cmake)
+include(source/ML/NNLS/sources.cmake)
 include(source/ML/sources.cmake)
 include(source/ML/SVM/sources.cmake)
 include(source/DATASTRUCTURES/sources.cmake)
@@ -57,7 +57,7 @@ include(include/OpenMS/INTERFACES/sources.cmake)
 include(include/OpenMS/CONCEPT/sources.cmake)
 include(include/OpenMS/SYSTEM/sources.cmake)
 include(include/OpenMS/MATH/MISC/sources.cmake)
-include(include/OpenMS/MATH/MISC/NNLS/sources.cmake)
+include(include/OpenMS/ML/NNLS/sources.cmake)
 include(include/OpenMS/MATH/STATISTICS/sources.cmake)
 include(include/OpenMS/ML/SVM/sources.cmake)
 include(include/OpenMS/DATASTRUCTURES/sources.cmake)
