+++ src/openms/source/ANALYSIS/QUANTITATION/IsobaricIsotopeCorrector.cpp
@@ -14,7 +14,7 @@
 #include <OpenMS/CONCEPT/LogStream.h>
 
 // NNLS isotope correction
-#include <OpenMS/MATH/MISC/NonNegativeLeastSquaresSolver.h>
+#include <OpenMS/ML/NNLS/NonnegativeLeastSquaresSolver.h>
 
 #include <Eigen/Core>
 #include <Eigen/LU>
