+++ src/openms/source/MATH/MISC/sources.cmake
@@ -10,7 +10,6 @@ EmgGradientDescent.cpp
 GridSearch.cpp
 LinearInterpolation.cpp
 MathFunctions.cpp
-NonNegativeLeastSquaresSolver.cpp
 MSNumpress.cpp
 RANSAC.cpp
 RANSACModel.cpp
