+++ src/tests/class_tests/openms/source/NNLS_test.cpp
@@ -10,7 +10,7 @@
 #include <OpenMS/test_config.h>
 
 ///////////////////////////
-#include <OpenMS/MATH/MISC/NNLS/NNLS.h>
+#include <OpenMS/ML/NNLS/NNLS.h>
 ///////////////////////////
 
 using namespace OpenMS;
